backend/requirements.txt
clic
Flask
gunicorn
itsdangerous
Jinja2
MarkupSafe
Werkzeug
varname
psycopg2
Flask-SQLAlchemy
Flask-Migrate
flask-cors
